Windows: Reparse Policy vs DFSLinks
When a reparse policy is specified and AFSLocateNameEntry() returns
with STATUS_REPARSE, do not re-evaluate the path with the reparse
policy disabled.  STATUS_REPARSE was returned because the FileObject's
FileName was modified and the IO Manager needs to reparse the request.

Reviewed-on: http://gerrit.openafs.org/11397
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>
(cherry picked from commit d36b8ff8db63ee893941d57422e54ca05e4a1564)

Change-Id: I740989489a4e0c268a033c4558c6a71b7b4ee3b9
Reviewed-on: http://gerrit.openafs.org/11409
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>